backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
motor==3.1.1
pymongo==4.4.0
pydantic>=1.10.0,<2.0.0
bcrypt==3.2.0
python-jose==3.3.0
python-dotenv==1.0.0
email-validator==2.0.0
requests==2.31.0
